feat(api): implemented auth endpoints, jwttokens, user config

+++ models/auth.go
@@ -15,7 +15,7 @@ type Session struct {
 
 	// UserID is the foreign key to the user
 	// @Description ID of the user this session belongs to
-	UserID uint `json:"userId" gorm:"not null"`
+	UserID uint64 `json:"userId" gorm:"not null"`
 
 	// RefreshToken is used to generate new access tokens
 	// @Description Refresh token value (hashed in database)
@@ -71,7 +71,7 @@ type TokenDetails struct {
 // JWTClaim defines the structure of the JWT claim
 // @Description JWT claim st eructure
 type JWTClaim struct {
-	UserID uint   `json:"userId"`
+	UserID uint64 `json:"userId"`
 	UUID   string `json:"uuid"`
 	Role   string `json:"role"`
 	jwt.RegisteredClaims
